Show message in the thread after sending
