+++ nextjs/serializers/thread.ts
@@ -1,24 +1,7 @@
-import { MentionsWithUsers } from '../types/apiResponses/threads/[threadId]';
 import { MessageWithAuthor } from '../types/partialTypes';
-import { SerializedAttachment, SerializedReaction } from '../types/shared';
-import type {
-  channels,
-  threads,
-  users,
-  messageAttachments,
-  messageReactions,
-} from '@prisma/client';
+import serializeMessage, { SerializedMessage } from './message';
 
-export interface SerializedMessage {
-  id: string;
-  body: string;
-  sentAt: string;
-  author?: users;
-  usersId: string;
-  mentions: MentionsWithUsers[];
-  attachments: SerializedAttachment[];
-  reactions: SerializedReaction[];
-}
+import type { channels, threads } from '@prisma/client';
 
 interface SerializedChannel {
   channelName: string;
@@ -49,43 +32,13 @@ function serializeChannel(channel?: channels): SerializedChannel | null {
   };
 }
 
-function serializeMessages(messages?: MessageWithAuthor[]) {
+function serializeMessages(
+  messages?: MessageWithAuthor[]
+): SerializedMessage[] {
   if (!messages) {
     return [];
   }
-  return messages.map((message: MessageWithAuthor) => {
-    return {
-      id: message.id,
-      body: message.body,
-      // Have to convert to string b/c Nextjs doesn't support date hydration -
-      // see: https://github.com/vercel/next.js/discussions/11498
-      sentAt: message.sentAt.toString(),
-      author: message.author,
-      usersId: message.usersId,
-      mentions: message.mentions || [],
-      attachments:
-        message.attachments
-          ?.map((attachment: messageAttachments) => {
-            return {
-              url: attachment.internalUrl,
-              name: attachment.name,
-            };
-          })
-          .filter(({ url }: SerializedAttachment) => Boolean(url)) || [],
-      reactions:
-        message.reactions
-          ?.filter(
-            (reaction: messageReactions) =>
-              typeof reaction.count === 'number' && reaction.count > 0
-          )
-          ?.map((reaction: messageReactions) => {
-            return {
-              type: reaction.name,
-              count: reaction.count,
-            } as SerializedReaction;
-          }) || [],
-    };
-  });
+  return messages.map(serializeMessage);
 }
 
 export default function serialize(
