+++ nextjs/utilities/featureFlags.ts
@@ -2,7 +2,7 @@ export const isStripeEnabled = false;
 export const isVercelDomainEnabled = true;
 export const isLoginProtectionEnabled = true;
 export const isFeedEnabled = false;
-export const isSendMessageEnabled = true;
+export const isSendMessageEnabled = false;
 
 export default {
   isStripeEnabled,
